마지막으로 리팩토링 하기

+++ src/logic/api/delete.ts
@@ -6,11 +6,17 @@ function deleteTodo(id: string) {
             Authorization: 'login token'
         }
     });
-    deleteItem.then(res => {
-        return res;
-    }).catch(err => {
-        console.log(err);
-    })
+    if (window.confirm('정말로 삭제하시겠습니까?') === true) {
+        alert('삭제되었습니다!');
+        deleteItem.then(res => {
+            window.location.reload();
+            return res;
+        }).catch(err => {
+            console.log(err);
+        })   
+    } else {
+        return;
+    }
 }
 
 export { deleteTodo };
